requirements.txt
torch
torchvision
numpy
Pillow
scikit-learn
matplotlib
seaborn
tensorboard
